+++ lib/acpi/acpi_table.c
@@ -122,12 +122,12 @@ void acpi_fill_header(struct acpi_table_header *header, char *signature)
 
 void acpi_align(struct acpi_ctx *ctx)
 {
-	ctx->current = (void *)ALIGN((ulong)ctx->current, 16);
+	ctx->current = (void *)ALIGN((uintptr_t)ctx->current, 16);
 }
 
 void acpi_align64(struct acpi_ctx *ctx)
 {
-	ctx->current = (void *)ALIGN((ulong)ctx->current, 64);
+	ctx->current = (void *)ALIGN((uintptr_t)ctx->current, 64);
 }
 
 void acpi_inc(struct acpi_ctx *ctx, uint amount)
